+++ script/build.ps1
@@ -13,7 +13,6 @@ $GIT_SHA = git rev-parse --short HEAD
 $VERSION = git rev-parse --abbrev-ref HEAD
 $LDFLAGS = "-X github.com/dremio/dremio-diagnostic-collector/v3/pkg/versions.GitSha=$GIT_SHA -X github.com/dremio/dremio-diagnostic-collector/v3/pkg/versions.Version=$VERSION"
 
-New-Item -ItemType File -Path .\cmd\root\ddcbinary\output\ddc.zip -Force
 # This assumes that you have 'go' installed in your environment
 $env:GOOS="linux"
 $env:GOARCH="amd64"
@@ -21,7 +20,17 @@ go build -ldflags "$LDFLAGS" -o .\bin\ddc .\cmd\local\main
 
 # Use Compress-Archive to create zip file and then move it
 Compress-Archive -Path .\bin\ddc -DestinationPath .\bin\ddc.zip
-Move-Item -Force -Path  .\bin\ddc.zip -Destination .\cmd\root\ddcbinary\output\ddc.zip
+Move-Item -Force -Path  .\bin\ddc.zip -Destination .\cmd\root\ddcbinary\output\ddc-amd64.zip
+Remove-Item -Path .\bin\ddc
+
+# This assumes that you have 'go' installed in your environment
+$env:GOOS="linux"
+$env:GOARCH="arm64"
+
+go build -ldflags "$LDFLAGS" -o .\bin\ddc .\cmd\local\main
+# Use Compress-Archive to create zip file and then move it
+Compress-Archive -Path .\bin\ddc -DestinationPath .\bin\ddc.zip
+Move-Item -Force -Path  .\bin\ddc.zip -Destination .\cmd\root\ddcbinary\output\ddc-arm64.zip
 Remove-Item -Path .\bin\ddc
 
 $env:GOOS="windows"