+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## [3.3.0] - UNRELEASED
+## [3.3.0] - 2024-01-03
 
 ### Added
 
 * capture container logs of all pods in namespace
+* cpu architecture detection for each host
+* The cpu architecture specific ddc local-collect binary is sent to the node
 
 ### Fixed